requirements.txt
aiocsv
aiofiles
beautifulsoup4
discord.py